backend/requirements.txt
Django==6.0.2
django-cors-headers==4.9.0
djangorestframework==3.17.1
djangorestframework-simplejwt>=5.3.1
openai==2.17.0
PyPDF2==3.0.1
python-docx==1.2.0
pydantic>=2.0.0
